+++ remoc/src/prelude.rs
@@ -55,4 +55,4 @@ pub use crate::rtc;
 
 #[cfg(feature = "rtc")]
 #[doc(no_inline)]
-pub use crate::rtc::{Client, Server, ServerRef, ServerRefMut, ServerShared, ServerSharedMut};
+pub use crate::rtc::{Client, ReqReceiver, Server, ServerRef, ServerRefMut, ServerShared, ServerSharedMut};