+++ remoc/tests/rtc/mod.rs
@@ -3,6 +3,7 @@ mod generics;
 mod readonly;
 mod simple;
 mod simple_clone;
+mod simple_req;
 mod value;
 
 // Must result in compile error: