+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to Remoc will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.14.0 - 2024-07-xx
+### Added
+- remote trait calling (RTC): generate ReqReceiver type for handling requests from
+  clients as messages
+
 ## 0.13.1 - 2024-07-14
 ### Added
 - codec: added Postcard codec by Firaenix